add code 5/6

+++ repository/UserRepository.go
@@ -29,6 +29,9 @@ func GetAllUsers(database db.Database) (*models.UserList, error) {
 	return list, nil
 }
 func AddUser(database db.Database, user *models.User) error {
+	if !isEmailValid(user.Email) {
+		return errors.New("email is wrong")
+	}
 	query := `INSERT INTO users (email) VALUES ($1)`
 	// err := database.Conn.QueryRow(query, user.Email)
 	// if err != nil {
@@ -45,15 +48,16 @@ func AddUser(database db.Database, user *models.User) error {
 	}
 	return nil
 }
-
-func GetUserByEmail(database db.Database, email string) (models.User, error){
+func GetUserByEmail(database db.Database, email string) (models.User, error) {
 	user := models.User{}
-
+	if !isEmailValid(email) {
+		return user, errors.New("email is wrong")
+	}
 	query := `select * from users where email = $1;`
 
 	err := database.Conn.QueryRow(query, email).Scan(&user.Email)
-	if err != nil{
-		if err == sql.ErrNoRows{
+	if err != nil {
+		if err == sql.ErrNoRows {
 			return user, err
 		}
 		return user, err
@@ -62,12 +66,15 @@ func GetUserByEmail(database db.Database, email string) (models.User, error){
 }
 
 func DeleteUser(database db.Database, email string) error {
+	if !isEmailValid(email) {
+		return errors.New("email is wrong")
+	}
 	query := `delete from users where email =$1`
 	_, err := database.Conn.Exec(query, email)
 	switch err {
-		case sql.ErrNoRows:
-			return db.ErrNoMatch
-		default:
-			return err
+	case sql.ErrNoRows:
+		return db.ErrNoMatch
+	default:
+		return err
 	}
 }
